simulator/requirements.txt
paho-mqtt==2.1.0
jsonschema==4.25.1
